# web/src/lib/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AssetFlow API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/categories": {
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "List Categories",
        "operationId": "list_categories_api_categories_get",
        "parameters": [
          {
            "name": "assetflow_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetflow Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Categories Api Categories Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "categories"
        ],
        "summary": "Add Category",
        "operationId": "add_category_api_categories_post",
        "parameters": [
          {
            "name": "assetflow_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetflow Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Add Category Api Categories Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/deposits": {
      "get": {
        "tags": [
          "deposits"
        ],
        "summary": "Get Deposits",
        "operationId": "get_deposits_api_deposits_get",
        "parameters": [
          {
            "name": "assetflow_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetflow Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepositSummary"
                  },
                  "title": "Response Get Deposits Api Deposits Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "deposits"
        ],
        "summary": "Add Deposit",
        "operationId": "add_deposit_api_deposits_post",
        "parameters": [
          {
            "name": "assetflow_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetflow Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/deposits/{deposit_id}": {
      "patch": {
        "tags": [
          "deposits"
        ],
        "summary": "Patch Deposit",
        "operationId": "patch_deposit_api_deposits__deposit_id__patch",
        "parameters": [
          {
            "name": "deposit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Deposit Id"
            }
          },
          {
            "name": "assetflow_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetflow Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "deposits"
        ],
        "summary": "Delete Deposit",
        "operationId": "delete_deposit_api_deposits__deposit_id__delete",
        "parameters": [
          {
            "name": "deposit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Deposit Id"
            }
          },
          {
            "name": "assetflow_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetflow Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reports/categories": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Category Dynamics",
        "operationId": "category_dynamics_api_reports_categories_get",
        "parameters": [
          {
            "name": "assetflow_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetflow Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDynamic"
                  },
                  "title": "Response Category Dynamics Api Reports Categories Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryDynamic": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "current_amount": {
            "type": "number",
            "title": "Current Amount"
          },
          "previous_amount": {
            "type": "number",
            "title": "Previous Amount"
          },
          "growth_pct": {
            "type": "number",
            "title": "Growth Pct"
          }
        },
        "type": "object",
        "required": [
          "category",
          "current_amount",
          "previous_amount",
          "growth_pct"
        ],
        "title": "CategoryDynamic"
      },
      "DepositCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "institution_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institution Name"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "principal_amount": {
            "type": "number",
            "title": "Principal Amount"
          },
          "current_balance": {
            "type": "number",
            "title": "Current Balance"
          },
          "annual_interest_rate": {
            "type": "number",
            "title": "Annual Interest Rate"
          },
          "payout_frequency": {
            "type": "string",
            "title": "Payout Frequency",
            "default": "monthly"
          },
          "capitalization_enabled": {
            "type": "boolean",
            "title": "Capitalization Enabled",
            "default": true
          },
          "opened_on": {
            "type": "string",
            "format": "date",
            "title": "Opened On"
          },
          "maturity_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maturity Date"
          },
          "next_payout_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Payout Date"
          },
          "early_withdrawal_terms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Early Withdrawal Terms"
          },
          "funding_account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Funding Account Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "currency",
          "principal_amount",
          "current_balance",
          "annual_interest_rate",
          "opened_on"
        ],
        "title": "DepositCreate"
      },
      "DepositSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "account_name": {
            "type": "string",
            "title": "Account Name"
          },
          "funding_account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Funding Account Id"
          },
          "funding_account_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Funding Account Name"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "institution_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institution Name"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "principal_amount": {
            "type": "number",
            "title": "Principal Amount"
          },
          "current_balance": {
            "type": "number",
            "title": "Current Balance"
          },
          "annual_interest_rate": {
            "type": "number",
            "title": "Annual Interest Rate"
          },
          "payout_frequency": {
            "type": "string",
            "title": "Payout Frequency"
          },
          "capitalization_enabled": {
            "type": "boolean",
            "title": "Capitalization Enabled"
          },
          "opened_on": {
            "type": "string",
            "format": "date",
            "title": "Opened On"
          },
          "maturity_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maturity Date"
          },
          "next_payout_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Payout Date"
          },
          "early_withdrawal_terms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Early Withdrawal Terms"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "account_id",
          "account_name",
          "name",
          "currency",
          "principal_amount",
          "current_balance",
          "annual_interest_rate",
          "payout_frequency",
          "capitalization_enabled",
          "opened_on",
          "status"
        ],
        "title": "DepositSummary"
      },
      "DepositUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "institution_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institution Name"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "principal_amount": {
            "type": "number",
            "title": "Principal Amount"
          },
          "current_balance": {
            "type": "number",
            "title": "Current Balance"
          },
          "annual_interest_rate": {
            "type": "number",
            "title": "Annual Interest Rate"
          },
          "payout_frequency": {
            "type": "string",
            "title": "Payout Frequency",
            "default": "monthly"
          },
          "capitalization_enabled": {
            "type": "boolean",
            "title": "Capitalization Enabled",
            "default": true
          },
          "opened_on": {
            "type": "string",
            "format": "date",
            "title": "Opened On"
          },
          "maturity_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maturity Date"
          },
          "next_payout_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Payout Date"
          },
          "early_withdrawal_terms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Early Withdrawal Terms"
          },
          "funding_account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Funding Account Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "open"
          }
        },
        "type": "object",
        "required": [
          "name",
          "currency",
          "principal_amount",
          "current_balance",
          "annual_interest_rate",
          "opened_on"
        ],
        "title": "DepositUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}